build: fix build
