+++ packages/core-browser/src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './electron';
+export * from './react-hooks';
 export * from './icon';
 export * from './json';
 export * from './label';
